requirements.txt
Flask==2.3.2
Flask-MySQL==1.5.2
Flask-WTF==1.1.1
Flask-Login==0.6.3
Werkzeug==2.3.6
Flask-Mail==0.9.1
wtforms==3.0.1